renamed index route to index
