+++ routes/index.js
@@ -1,0 +1,11 @@
+'use strict';
+
+const express = require('express');
+const router = express.Router();
+
+router.route('/')
+  .get((req, res, next) => {
+    res.render('index');
+  });
+
+module.exports = router;
